Fix nasty bug when cmake is run for the first time.
This bug cause the user set variables to be shadowed
by the cache defaults. the cache default of resources extra
path is an empty string and so it would glob the entire file
system. It would also set to CPU_BACKEND.

By setting this CMake policy we use NEW behaviour to not
reset non-cache vars when cache is set for the first time.