WIP: Add monkey_demo sample for Vulkan.
warning! this is very messy right now. it doesn't even work entirely.
I just wanted to get a commit for a working state.

it seems that vk dynamic rendering is "just broken" on my setup.
afaik, I am using the api correct, but no frag shader waves are launched.
literally changing nothing but to use vkrenderpass fixes.
I saw in renderdoc that that renderarea was cleared to all zero.
where in a working rdc trace, renderarea was nonzero. I think that this is
the root cause.