Fix the stuttering due to Sleep.
this function is not reliable. if we wanted to wait for 1ms
plus a few 1us, it could wait for many more 1us over 1ms.
hence, we don't use sleep when there is only a little time to
wait and we go right into the loop that burns cycles until
we hit the wall clock time we want.