Sort of fix the "input jitter" problem.
so it seems that there will ALWAYS be jitter. maybe.
I ran minecraft and observed the same sort of jitter
  when doing the same thing (walk orbitally around a
  thing and keep your first person camera looking at
  that thing).
It looks to just be what happens when you sample
  the game simulation at 60HZ.
if the function of camera movement + rotation
  are both periodic and a specific combination,
  the 60hz frame gives aliasing artifacts.

with this commit I have done some smoothing to
  the raw mouse input to make that less sporadic,
  and visually I observe less jitter.
I think things will appear just barely good enough
  if we add some motion blur pass to the mix.

and I don't know, maybe in the future there is some
anti aliasing idea hiding.