ssr wip: fix nasty bug.
so things aren't quite working well at all yet.
but we just fixed an issue with this one (sort of).
where, we before would falsely register a hit
in the case that the ray goes behind an object.

the idea is that we need to find the intersection
point. so, the delta between the depth buffer and
the ray depth should be small.

with this change however things are aggressive and
we are missing reflections.