WIP stuff.
+ fixed Platform_showMouse, where our higher-level API
doesn't map proper to the win32 one. our API is a toggle
whereas the win32 one is an accumulator.

+ added GameHandleInput. games now handle their input
in this callback. input is collected on the main thread, and
rendering has been split to a separate thread.
the main thread, goes much faster, so we collect and apply
input to the game simulation in a timely manner.

+ add "proper" smooth camera and movement for monkey
demo app. this added stuff isn't really physics correct but
it at least seems to have some of that good game feel.