decouple input handle loop from main message loop.
from within the main message loop,
we now PostMessage of only the input message kinds.
we post these to the hidden window associated with the input
handle loop.

this way, the input handle loop is never blocked and we can always
get it to run at that nice whatever high Hz we want.